+++ vpin-connector-iscored/src/main/java/de/mephisto/vpin/connectors/iscored/IScored.java
@@ -191,7 +191,7 @@ public class IScored {
       out.close();
       conn.disconnect();
 
-      LOG.info("Submitted new highscore to iscored game room \"" + gameRoom.getName() + "\"");
+      LOG.info("Submitted new highscore to iscored game room \"" + gameRoom.getName() + "\": name=" + name + ", game=" + game.getId() + ", score=" + highscore);
       return true;
     } catch (IOException e) {
       LOG.error("Failed to submit iscored highscore: " + e.getMessage(), e);