+++ RELEASE_NOTES.md
@@ -1,6 +1,10 @@
 ## Release Notes 3.0.3
 
 - **VPin Mania**: Optimized wheel icon lookup with a persistent cache, so that when browsing the tables by letter in VPin Mania everything is loaded faster.
+- **VPin Mania**: Fixed joining tournaments.
+- **VPin Mania**: Fixed wrong status of iScored checkbox "Score submissionss enabled" which was always unchecked for existing tournaments.
+- **VPin Mania**: Fixed duplicated entries issue in the players list.
+- 
 - **Table Overview / ALT Sound**: Fixed ALT sound editor not showing when "Edit" was pressed.
 - **Highscore Cards**: Added "Reset Default Asset" button to the default asset preview.