+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
   <img src="https://raw.githubusercontent.com/syd711/vpin-studio/main/documentation/preferences/vpxmonitor.png" width="700" />
 
+- Added separate "VPin Mania Name" field to the player edit dialog. This name will be used for all VPin Mania based stuff. You can rename your existing accounts too.
 
 ### Bugfixes
 
@@ -14,4 +15,5 @@
 - **VPin Mania**: Fixed filtering of tournament highscore submissions when no matching player has been found. 
 - **iScored**: Fixed filtering of highscores submission to iScored where on matching player has been found. 
 - **Table Overview / ALT Sound**: Fixed ALT sound editor not showing when "Edit" was pressed.
+- **Table Overview / VPS Updates**: Fixed filtering for VPS updates according to the settings.
 - **Highscore Cards**: Added "Reset Default Asset" button to the default asset preview.