+++ RELEASE_NOTES.md
@@ -1,3 +1,6 @@
-## Release Notes 3.0.4
+## Release Notes 3.0.5
 
-- **Server Startup**: Fixed critical bug where the emulator launch script updates did not work. As a result, the server did not detect any highscore changes after table exit. This happened only to new version 3 installations.
+**Bugfixes**
+
+- **Table Resolving**: The early date format error of the initial 3.0 release led to some corrupt data. The server has been made more robust that these tables are now loaded anyway (and repaired again on update).
+